feat: opt code
